requirements.txt
aws-cdk-lib==2.155.0
aws-cdk.core
constructs>=10.0.0,<11.0.0
